Refactoring of the stubs_generator package
