+++ src/safeds_stubgen/stubs_generator/__init__.py
@@ -1,8 +1,9 @@
 """API-Analyzer for the Safe-DS stubs generator."""
-
 from __future__ import annotations
 
-from ._generate_stubs import NamingConvention, StubsStringGenerator, create_stub_files, generate_stub_data
+from ._generate_stubs import create_stub_files, generate_stub_data
+from ._helper import NamingConvention
+from ._stub_string_generator import StubsStringGenerator
 
 __all__ = [
     "create_stub_files",
